Disable travis publishing the page when it succeeds if it's a PR. Fix a trivial review comment to test this works.

+++ 02stl/cpp/particlesMMap.cc
@@ -8,17 +8,17 @@ int main()
   FILE* ifp = fopen("02stl/cpp/particleList.txt","r");
 
   // Read in the data
-  std::multimap<std::string,float> theParticles;
+  std::multimap<std::string,double> theParticles;
   char name[80];
-  float momentum;
+  double momentum;
   while (!feof(ifp)) {
     fscanf(ifp, "%s %f", name, &momentum);
-    if (!feof(ifp)) theParticles.insert( std::pair<std::string,float>(std::string(name),momentum) );
+    if (!feof(ifp)) theParticles.insert( std::pair<std::string,double>(std::string(name),momentum) );
   }
   fclose(ifp);
 
   // Output - it's already sorted!
-  std::multimap<std::string,float>::iterator iter = theParticles.begin();
+  std::multimap<std::string,double>::iterator iter = theParticles.begin();
   for ( ; iter!=theParticles.end(); ++iter) {
     std::cout << (*iter).first << " " << (*iter).second << std::endl;
   }
