+++ 02stl/cpp/particlesMap.cc
@@ -4,7 +4,7 @@
 #include <iostream>
 
 /// func
-void keepMax(std::map<std::string,float>& theMap, const std::string key, const float value)
+void keepMax(std::map<std::string,double>& theMap, const std::string key, const double value)
 {
   if (theMap.find(key)==theMap.end()) //element doesn't already exist in map
     theMap[key] = value;
@@ -17,12 +17,12 @@ void keepMax(std::map<std::string,float>& theMap, const std::string key, const f
 class maxMap
 {
 public:
-  typedef std::map<std::string,float>::iterator iterator;
+  typedef std::map<std::string,double>::iterator iterator;
   
-  void insert(const std::pair<std::string,float> theElement)
+  void insert(const std::pair<std::string,double> theElement)
   {
     std::string key = theElement.first;
-    float value = theElement.second;
+    double value = theElement.second;
     if (m_map.find(key)==m_map.end()) //element doesn't already exist in map
       m_map[key] = value;
     else
@@ -32,7 +32,7 @@ public:
   iterator begin() { return m_map.begin(); }
   iterator end() { return m_map.end(); }
 private:
-  std::map<std::string,float> m_map;
+  std::map<std::string,double> m_map;
 };
 
 
@@ -44,10 +44,10 @@ int main()
   // Read in the data
   maxMap theParticles;
   char name[80];
-  float momentum;
+  double momentum;
   while (!feof(ifp)) {
     fscanf(ifp, "%s %f", name, &momentum);
-    if (!feof(ifp)) theParticles.insert( std::pair<std::string,float>(std::string(name),momentum) );
+    if (!feof(ifp)) theParticles.insert( std::pair<std::string,double>(std::string(name),momentum) );
   }
   fclose(ifp);
